requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-rest==0.8.7
djangorestframework==3.14.0
Pillow==9.2.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
